Added updating logic.
The script will check for updates once a day, plus there is a new menu item for forcing an update check. If an update is found, it will prompt the user to install the new version. Most of this comes from https://github.com/Benjol/SE-AutoReviewComments/blob/master/dist/autoreviewcomments.user.js, with a few upgrades.