Version 1.4; Cross-site compatibility; Removed usage info; Short reasons; Updater updated; Restructuring; Misc Bugfixes
This script should now work on all Stack Exchange sites, please open an issue if you find any that it doesn't work on.

All of the relevant usage information is in README.md and I've changed the `@namespace` to point to the github url, that information is no longer necessary to have in the script itself.

In response to a feature request made in chat, if you enter one letter into the reason dialog and submit it, if that letter matches a predefined reason, it will expand the reason.

> @TinyGiant I'd love to choose a reason in a single keystroke /durron597   
> <sup>http://chat.stackoverflow.com/transcript/message/25121337#25121337</sup>

I've simplified the updater to use the scripts `@version` instead of a separate variable. 

I grouped the functions and did some other general housekeeping.

Other miscellaneous bugfixes.