Use "raw" URL for SO Close Vote Request Generator, not "blob"
Using the "blob" URL requires clicking on the "raw" button on GitHub as an additional step after visiting the link, in order to get [Grease|Tamper|...]monkey to ask to install the script. Using the "raw" URL makes the behaviour consistent with the description in the surrounding text.

The "raw" URL is also used for the MagicEditor further down.