Fix some deleted messages not showing up in transcript
The code to search for deleted messages in transcripts worked by grabbing 2 pages (1000 events) from the target chat room in reverse order, starting 15000 message IDs past the end of the transcript. This can fail for high-traffic rooms such as CHQ, if the room's chat traffic accounts for more than 1/15 of chat messages on the server.

This PR fixes that issue by continuing to fetch events until we reach the beginning of the visible transcript, instead of stopping after two pages.