Updated the classname for the top bar element
With the most recent SE deploy (seems this change was globally applied, on all SE sites), the top bar element has the class `s-topbar` instead of the previous `top-bar`, so this change broke the *-pls userscript.
The change here updates the class name in the script to be inline with the current state of the SE sites.